Move components to it's own subfolders
