+++ src/components/ContentLayout/ContentLayout.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 
 import Content from "./Content";
-import Header from "./Header";
+import Header from "../Header/Header";
 
 import "./ContentLayout.css";
